refactor(BTC - Fees): hard fail if fees endpoint is down
