+++ packages/blockchain-wallet-v4/src/redux/payment/model.js
@@ -1,0 +1,1 @@
+export const FETCH_FEES_FAILURE = 'fetch_fees_failure'
